Fixed warnings causing ns3 compile to fail.

+++ ns-allinone-3.28/ns-3.28/src/adhoc-routing/helper/adhoc-routing-helper.cc
@@ -159,7 +159,7 @@ void AdHocRoutingHelper::updateLinkBandwidth(uint32_t bandwidthBytes)
     Simulator::Schedule(Seconds(1.0), &RemovePacketFromBandwidthMetric, this, numberOfBits);
 }
 
-uint32_t AdHocRoutingHelper::getNs3SimulatedTimeMS()
+uint64_t AdHocRoutingHelper::getNs3SimulatedTimeMS()
 {
     return (uint32_t)Simulator::Now().GetMilliSeconds();
 }
